Update calculation for K_scale in fused_rope_and_quant
before, last step of scale calculation, /2**k_bit-1, was done in
triton, this cause it to yield different result from the reference,
where this step was done in python.
Now I moved the step outside the triton kernel, in the function
after triton was called. And the result is the same as the
reference.