FIX: fused_rope_and_quant overwrot memory when storing k_{mn, scale}
When storing k_mn and k_scale, the mask in tl.store was not right,
it was not checking the k_mn and scale boundaries correctly.
This causes some blocks to write outside k_mn and k_scale's
allocated memory, making example.py get wrong answers.

The updated mask checks so that the sid (sequence group) is less
than the possible group of sequences. And when sid exceeds groups,
this value is not stored.

This might have happened because the kernel was launched with
tl.cdiv(seqlen, group_size), which take the ceiling to launch enough
blocks, but we take the floor when using k_mn and scale.

What I learnt:
- Triton does not have error check against memory overlap, if the
	memory is allocated.
- how to manipulate mask to avoid overwriting memory